+++ src/main/java/fi/helsinki/btls/mqtt/UbiMqttProvider.java
@@ -1,4 +1,4 @@
-package fi.helsinki.btls;
+package fi.helsinki.btls.mqtt;
 
 import org.eclipse.paho.client.mqttv3.IMqttToken;
 import fi.helsinki.ubimqtt.IUbiActionListener;
