+++ src/main/java/fi/helsinki/btls/mqtt/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * mqtt package.
+ */
+package fi.helsinki.btls.mqtt;
