+++ src/main/java/fi/helsinki/btls/trilaterion/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * trilaterion page.
+ */
+package fi.helsinki.btls.trilaterion;
