+++ src/main/java/fi/helsinki/btls/utils/ObserverService.java
@@ -1,4 +1,4 @@
-package fi.helsinki.btls;
+package fi.helsinki.btls.utils;
 
 import java.util.ArrayList;
 import java.util.HashMap;
