+++ src/main/java/fi/helsinki/btls/utils/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * utils package.
+ */
+package fi.helsinki.btls.utils;
