+++ src/test/java/fi/helsinki/btls/datamodels/BeaconTest.java
@@ -1,7 +1,6 @@
 package fi.helsinki.btls.datamodels;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 import org.junit.Before;
 import org.junit.Test;
 import java.time.LocalDateTime;
