+++ src/test/java/fi/helsinki/btls/trilaterion/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * trilaterion test package.
+ */
+package fi.helsinki.btls.trilaterion;
